% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlightHTMLmaster.r
\name{highlight_html}
\alias{highlight_html}
\title{Master highlight HTML function}
\usage{
highlight_html(input, output, tags, update_css = TRUE, browse = TRUE,
  print = FALSE)
}
\arguments{
\item{input}{File name of HTML file to highlight the cells of the table or text}

\item{output}{Output file name of highlighted HTML file}

\item{tags}{character vector with CSS tags to be added}

\item{update_css}{TRUE/FALSE variable indicating whether the CSS should be updated}

\item{browse}{logical, If TRUE (default) output file opens in default browser, if FALSE, 
file is written, but not opened in browser.}

\item{print}{logical, if TRUE print output to R console, if false (default) output is 
filtered to other methods (see browse or output).}
}
\description{
This function inputs and exports an HTML file.  The HTML file is then processed
to search for tags that inject CSS automatically into the HTML file.
}
\details{
A function that allows the alteration of HTML using CSS.  This may be helpful
coming from a markdown or R markdown file to alter aspects of the page based on 
a specific criteria.  This function handles both tables as well as normal text.  
The options are only limited based on your knowledge of CSS.
}
\examples{
\donttest{
# Example of simple test table
# Change background color of table cells
library(highlightHTML)

# Setting path for example html files 
# To see path where these are saved, type file or file1 in the 
# r console.
file <- system.file('examples', 'bgtable.html', package = 'highlightHTML')

# Creating CSS tags to inject into HTML document
tags <- c("#bgred {background-color: #FF0000;}", 
  "#bgblue {background-color: #0000FF;}")
  
# Command to post-process HTML file - Writes to temporary file
highlight_html(input = file, output = tempfile(fileext = ".html"),  
  tags = tags, update_css = TRUE, browse = FALSE)
  
# Change background color and text color with CSS
tags <- c("#bgred {background-color: #FF0000; color: white;}", 
  "#bgblue {background-color: #0000FF; color: white;}")
  
# Post-process HTML file
highlight_html(input = file, output = tempfile(fileext = ".html"),  
  tags = tags, update_css = TRUE, browse = TRUE, print = FALSE)

# By default the new file is opened in your default browser, here set to FALSE
highlight_html(input = file, output = tempfile(fileext = ".html"), 
  tags = tags, update_css = TRUE, browse = FALSE, print = FALSE)
  
# Setting path for example html files 
# To see path where these are saved, type file or file1 in the 
# r console.
file <- system.file('examples', 'bgtext.html', package = 'highlightHTML')

# Change background color and text color with CSS
tags <- c("#bgblack {background-color: black; color: white;}", 
  "#bgblue {background-color: #0000FF; color: white;}",
  "#colgreen {color: green;}")
  
# Post-process HTML file
highlight_html(input = file, output = tempfile(fileext = ".html"), 
  tags = tags, update_css = TRUE, browse = TRUE)
}
  
}

