% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlightHTML.r
\name{highlight_html_cells}
\alias{highlight_html_cells}
\title{Highlights HTML table cells}
\usage{
highlight_html_cells(x, output, tags, update_css = TRUE, browse = TRUE,
  print = FALSE)
}
\arguments{
\item{x}{File name of HTML file to highlight the cells of the table}

\item{output}{Output file name of highlighted HTML file}

\item{tags}{character vector with CSS tags to be added}

\item{update_css}{TRUE/FALSE variable indicating whether the CSS should be updated.}

\item{browse}{logical, If TRUE (default) output file opens in default browser, if FALSE, 
file is written, but not opened in browser.}

\item{print}{logical, if TRUE print output to R console, if false (default) output is 
filtered to other methods (see browse or output).}
}
\description{
Inputs an HTML file and outputs a new HTML file that adds css to highlight background of table cells.
}
\details{
A function that allows the alteration of cells in an HTML table using CSS.  This may be helpful
coming from a markdown or R markdown file to highlight certain cells based on a specific criteria,
such as values greater than 85 should be highlighted as blue or text color changed to green.
}

