% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlightHTML.r
\name{highlight_html_text}
\alias{highlight_html_text}
\title{Highlights HTML text}
\usage{
highlight_html_text(x, output, tags, update_css = TRUE, browse = TRUE,
  print = FALSE)
}
\arguments{
\item{x}{File name of HTML file to highlight the cells of the table}

\item{output}{Output file name of highlighted HTML file}

\item{tags}{character vector with CSS tags to be added}

\item{update_css}{TRUE/FALSE variable indicating whether the CSS should be updated}

\item{browse}{logical, If TRUE (default) output file opens in default browser, if FALSE, 
file is written, but not opened in browser.}

\item{print}{locical, if TRUE print output to R console, if false (default) output is 
filtered to other methods (see browse or output).}
}
\description{
Inputs a HTML file and outputs the same HTML file, but now with injected CSS to 
change the appearance of text.
}
\details{
A function that allows the alteration of text in a HTML file using CSS. This can be most helpful
when converting a markdown file to a HTML file to display as a webpage.  The function 
will post process the HTML file to add additional formatting without editing the raw HTML file.
}

