% Generated by roxygen2 (4.0.1): do not edit by hand
\name{highlightHTML}
\alias{highlightHTML}
\title{Master highlight HTML function}
\usage{
highlightHTML(input, output, updateCSS = TRUE, tags, browse = TRUE,
  print = FALSE)
}
\arguments{
\item{input}{File name of HTML file to highlight the cells of the table}

\item{output}{Output file name of highlighted HTML file}

\item{updateCSS}{TRUE/FALSE variable indicating whether the CSS should be updated.}

\item{tags}{character vector with CSS tags to be added}

\item{browse}{logical, If TRUE (default) output file opens in default browser, if FALSE,
file is written, but not opened in browser.}

\item{print}{logical, if TRUE print output to R console, if false (default) output is
   filtered to other methods (see browse or output).}
}
\description{
This function inputs and exports an HTML file.  The HTML file is then processed
to search for tags that inject CSS automatically into the HTML file.
}
\details{
A function that allows the alteration of HTML using CSS.  This may be helpful
coming from a markdown or R markdown file to alter aspects of the page based on
a specific criteria.  This function handles both tables as well as normal text.
The options are only limited based on your knowledge of CSS.
}

