% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.highlightHTML.r
\name{print.highlightHTML}
\alias{print.highlightHTML}
\title{Prints highlightHTML object}
\usage{
\method{print}{highlightHTML}(x, ...)
}
\arguments{
\item{x}{An object from highlightHTML function.}

\item{...}{Additional arguments passed to function.}
}
\description{
Prints highlightHTML object
}
