% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rgb2hex}
\alias{rgb2hex}
\title{Convert RGB to hex}
\usage{
rgb2hex(rgbcode = NULL, rcolor = NULL)
}
\arguments{
\item{rgbcode}{List of rgb color codes, each list must be a vector of three objects representing
the three components of rgb color code from 0 - 255.}

\item{rcolor}{A list of R color names}
}
\description{
Enter a list of RGB color codes, or R colors, and get the appropriate hex color code.
}
\examples{
rgb2hex(rcolor = list("sienna2", "thistle1"))
rgb2hex(rcolor = list("sienna2", "thistle1"), rgbcode = list('orange' = c(238, 74, 24),
'raw umber' = c(113, 75, 35)))
rgb2hex(rgbcode = list('orange' = c(238, 74, 24), 'raw umber' = c(113, 75, 35)))
}

