% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyApp.r
\name{shinyRGB2hex}
\alias{shinyRGB2hex}
\title{Run shiny app}
\usage{
shinyRGB2hex(...)
}
\arguments{
\item{...}{Other arguments to pass, currently does nothing.}
}
\description{
Function that automatically opens shiny app to convert rgb codes to hexidecimal codes
}
