% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlightHTML.r
\name{update_css}
\alias{update_css}
\title{Update external CSS file}
\usage{
update_css(x, tags)
}
\arguments{
\item{x}{Name of the R object after readLines}

\item{tags}{Character vector of tags to add to CSS}
}
\description{
Update external CSS file
}

