% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlightHTML.r
\name{updateCSS}
\alias{updateCSS}
\title{Update external CSS file}
\usage{
updateCSS(input, tags)
}
\arguments{
\item{input}{Name of the R object after readLines}

\item{tags}{Character vector of tags to add to CSS}
}
\description{
Update external CSS file
}

