% Generated by roxygen2 (4.0.1): do not edit by hand
\name{highlightHTMLcells}
\alias{highlightHTMLcells}
\title{Highlights HTML table cells}
\usage{
highlightHTMLcells(input, output, updateCSS = TRUE, tags, browse = TRUE,
  print = FALSE)
}
\arguments{
\item{input}{File name of HTML file to highlight the cells of the table}

\item{output}{Output file name of highlighted HTML file}

\item{updateCSS}{TRUE/FALSE variable indicating whether the CSS should be updated.}

\item{tags}{character vector with CSS tags to be added}

\item{browse}{logical, If TRUE (default) output file opens in default browser, if FALSE,
file is written, but not opened in browser.}

\item{print}{locical, if TRUE print output to R console, if false (default) output is
filtered to other methods (see browse or output).}
}
\description{
Inputs an HTML file and outputs a new HTML file that adds css to highlight background of table cells.
}
\details{
A function that allows the alteration of cells in an HTML table using CSS.  This may be helpful
coming from a markdown or R markdown file to highlight certain cells based on a specific criteria,
such as values greater than 85 should be highlighted as blue or text color changed to green.
}
\examples{
# Example of simple test table
# Change background color of table cells
require(highlightHTML)

# Setting path for example html files
# To see path where these are saved, type file or file1 in the
# r console.
file <- system.file('examples', package = 'highlightHTML')
file1 <- paste(file, "bgtable.html", sep = "/")

# Creating CSS tags to inject into HTML document
tags <- c("#bgred {background-color: #FF0000;}",
  "#bgblue {background-color: #0000FF;}")

# Command to post-process HTML file - Writes to temporary file
highlightHTMLcells(input = file1, output = tempfile(fileext = ".html"), updateCSS = TRUE,
  tags = tags, browse = FALSE)

# Change background color and text color with CSS
tags <- c("#bgred {background-color: #FF0000; color: white;}",
  "#bgblue {background-color: #0000FF; color: white;}")

# Post-process HTML file
highlightHTMLcells(input = file1, output = NULL, updateCSS = TRUE,
  tags = tags, browse = TRUE)

# By default the new file is opened in your default browser, here set to FALSE
highlightHTMLcells(input = file1, output = tempfile(fileext = ".html"), updateCSS = TRUE,
  tags = tags, browse=FALSE)
}

