% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_inject.r
\name{table_id_inject}
\alias{table_id_inject}
\title{Table hash addition}
\usage{
table_id_inject(x, id, conditions, variable = NULL)
}
\arguments{
\item{x}{A table object}

\item{id}{A vector of css id(s) to include}

\item{conditions}{A vector of conditions to include id. Must be same length as id}

\item{variable}{An optional vector of column names to specify search of conditions.}
}
\description{
A helper function to include a hashtag id code within an R table.
}

